requirements.txt
Flask==3.1.2
Flask_Minify==0.50
flask_session
gunicorn
Pillow==12.1.0
git+https://github.com/notashur/modified-kurigram.git
python-dotenv==1.2.1
pywebpush==2.0.3
Requests==2.32.5
TgCrypto